client: Split ArtifactsList & SpaceshipsList
